requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
boltons==23.0.0
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
Django==4.1.7
django-computedfields==0.2.2
django-enumchoicefield==3.0.1
django-enumfields==2.1.1
django-extensions==3.2.1
django-fast-update==0.2.3
django-filter==23.2
django-intenumfield==1.4.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
jsonpatch==1.32
jsonpointer==2.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
pip==23.0.1
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycosat==0.6.3
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
setuptools==67.8.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
wheel==0.38.4
yarl==1.8.2
zstandard==0.19.0
